Add files via upload
Home Automation using IOT provides flexibility to control the electrical appliances in the house. Home Automation using IOT provides flexibility to control the electrical appliances in the house. Home Automation is meant for home and the technological solution that automates the bulk of electrical and the technology-based tasks within the home. It uses a combination of hardware and software technologies that is used to enable the control and the management over appliances and devices within the home. Home automation is also known as domotics (is the encounter of information technology, electro-techniques and electronics that makes a home become "smart") and the home with an automation system is also known as a smart home. Home Automation using IOT (Internet of Things) is nothing but it provides flexibility to control the electrical appliances in the house including bulb, ceiling fan, TV, music system, etc through your smart phone. Home Automation has various advantages like comfort, increased security, and energy efficiency. The primary motive of this project is that helps the old and the handicapped people live a more independent life. The objective of this Home Automation is to take into consideration all the domestic systems that are difficult to control by the old and the handicapped people. The proposed system allows the users to have access to all the electrical appliances in the house including bulb, ceiling fan, TV, music system, etc with a single click on a smart phone to turn it either ON or OFF via Bluetooth. The most important is that it has to be user friendly and simple to operate.